git-svn-id: https://clinkc.svn.sourceforge.net/svnroot/clinkc/trunk@401 bfd3be0d-9f21-0410-925f-e30e9767a5cc

+++ wrapper/objc/CyberLink/CGUpnpControlPoint.m
@@ -72,18 +72,18 @@
 		cg_sleep(mx * 1000);
 }
 
-- (int)ssdpSearchMX
+- (NSInteger)ssdpSearchMX
 {
 	if (!cObject)
 		return 0;
 	return cg_upnp_controlpoint_getssdpsearchmx(cObject);
 }
 
-- (void)setSsdpSearchMX:(int)mx;
+- (void)setSsdpSearchMX:(NSInteger)mx;
 {
 	if (!cObject)
 		return;
-	cg_upnp_controlpoint_setssdpsearchmx(cObject, mx);
+	cg_upnp_controlpoint_setssdpsearchmx(cObject, (int)mx);
 }
 
 - (NSArray *)devices
